Added validation for update, delete and get methods

+++ CustomerAPI/src/main/java/com/test/CustomerAPI/service/CustomerCrudService.java
@@ -1,15 +1,14 @@
 package com.test.CustomerAPI.service;
 
 import com.test.CustomerAPI.Entity.Customer;
-
+import com.test.CustomerAPI.exception.CustomerNotFoundException;
 import com.test.CustomerAPI.exception.DuplicateEmailException;
 import com.test.CustomerAPI.exception.InvalidCustomerDataException;
 import com.test.CustomerAPI.repository.CustomerRepository;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import java.util.*;
+import java.util.List;
+import java.util.UUID;
 
 @Service
 public class CustomerCrudService implements CustomerService {
@@ -28,7 +27,6 @@ public class CustomerCrudService implements CustomerService {
     public List<Customer> findAllCustomers() {
 
         List<Customer> allCustomers = customerRepository.findAll();
-
         return allCustomers;
     }
 
@@ -36,12 +34,14 @@ public class CustomerCrudService implements CustomerService {
     // Find customer by their id
     @Override
     public Customer findCustomerById(UUID id) {
-        Optional<Customer> customerDetails = customerRepository.findById(id);
 
-        // Handle the Optional result here
-        return customerDetails.orElseThrow(() -> new NoSuchElementException("Customer not found with ID: " + id));
+        Customer customerDetails =  customerRepository.findById(id)
+                .orElseThrow(() -> {
+                    return new CustomerNotFoundException("Customer with ID " + id + " does not exist");
+                });
+
+        return customerDetails;
     }
-
 
 
     // Creation of a new customer
@@ -78,40 +78,30 @@ public class CustomerCrudService implements CustomerService {
     }
 
 
-
     // Updating details for an existing customer
     @Override
     public Customer updateCustomer(UUID id, Customer customerDetails) {
-        Optional<Customer> existingCustomerOptional = customerRepository.findById(id);
 
-        if (existingCustomerOptional.isPresent()) {
-            Customer existingCustomer = existingCustomerOptional.get(); // Retrieve the actual Customer object
+        Customer existingCustomer = customerRepository.findById(id)
+                .orElseThrow(() -> new CustomerNotFoundException("Customer with ID " + id + " does not exist"));
 
-            // Update the fields
-            existingCustomer.setFirstName(customerDetails.getFirstName());
-            existingCustomer.setMiddleName(customerDetails.getMiddleName());
-            existingCustomer.setLastName(customerDetails.getLastName());
-            existingCustomer.setEmailAddress(customerDetails.getEmailAddress());
-            existingCustomer.setPhoneNumber(customerDetails.getPhoneNumber());
 
-            // Save the updated customer back to the repository
-            return customerRepository.save(existingCustomer);
-        } else {
-            throw new NoSuchElementException("Customer not found with ID: " + id);
-        }
+        existingCustomer.setFirstName(customerDetails.getFirstName());
+        existingCustomer.setMiddleName(customerDetails.getMiddleName());
+        existingCustomer.setLastName(customerDetails.getLastName());
+        existingCustomer.setEmailAddress(customerDetails.getEmailAddress());
+        existingCustomer.setPhoneNumber(customerDetails.getPhoneNumber());
+        Customer updatedCustomer = customerRepository.save(existingCustomer);
+        return updatedCustomer;
     }
-
-
-
-
-
-
 
 
     // Delete customer
     @Override
     public void deleteCustomer(UUID id) {
-
+        if (!customerRepository.existsById(id)) {
+            throw new CustomerNotFoundException("Customer with ID " + id + " does not exist");
+        }
 
         customerRepository.deleteById(id);
